CLEANUP: Remove unused variables in loops/external.c

+++ src/ViennaRNA/loops/external.c
@@ -550,7 +550,7 @@ decompose_f3_ext_stem_d0(vrna_fold_compound_t       *fc,
   if (length <= i + maxdist)
     e = MIN2(e, stems[length]);
 
-  stems += i;
+  stems += (size_t)i;
   free(stems);
 
   return e;
@@ -600,7 +600,7 @@ decompose_f3_ext_stem_d2(vrna_fold_compound_t       *fc,
   if (length <= i + maxdist)
     e = MIN2(e, stems[length]);
 
-  stems += i;
+  stems += (size_t)i;
   free(stems);
 
   return e;
@@ -699,7 +699,7 @@ decompose_f3_ext_stem_d1(vrna_fold_compound_t       *fc,
   if (length <= i + maxdist)
     ee = MIN2(ee, stems[length]);
 
-  stems += i;
+  stems += (size_t)i;
   free(stems);
 
   e = MIN2(e, ee);
@@ -714,7 +714,7 @@ decompose_f3_ext_stem_d1(vrna_fold_compound_t       *fc,
   if (length <= i + maxdist)
     ee = MIN2(ee, stems[length]);
 
-  stems += i;
+  stems += (size_t)i;
   free(stems);
 
   e = MIN2(e, ee);
@@ -729,7 +729,7 @@ decompose_f3_ext_stem_d1(vrna_fold_compound_t       *fc,
   if (length <= i + maxdist)
     ee = MIN2(ee, stems[length]);
 
-  stems += i;
+  stems += (size_t)i;
   free(stems);
 
   e = MIN2(e, ee);
@@ -744,7 +744,7 @@ decompose_f3_ext_stem_d1(vrna_fold_compound_t       *fc,
   if (length <= i + maxdist)
     ee = MIN2(ee, stems[length]);
 
-  stems += i;
+  stems += (size_t)i;
   free(stems);
 
   e = MIN2(e, ee);
@@ -998,7 +998,7 @@ f3_get_stem_contributions_d0(vrna_fold_compound_t       *fc,
                      (zsc_data->pre_filter)) ? 1 : 0;
 #endif
   stems = (int *)vrna_alloc(sizeof(int) * (maxdist + 6));
-  stems -= i;
+  stems -= (size_t)i;
 
   sc_spl_stem = sc_wrapper->decomp_stem;
   sc_red_stem = sc_wrapper->red_stem;
@@ -1291,7 +1291,7 @@ f3_get_stem_contributions_d2(vrna_fold_compound_t       *fc,
 #endif
 
   stems = (int *)vrna_alloc(sizeof(int) * (maxdist + 6));
-  stems -= i;
+  stems -= (size_t)i;
 
   sc_spl_stem = sc_wrapper->decomp_stem;
   sc_red_stem = sc_wrapper->red_stem;
@@ -1592,7 +1592,7 @@ f3_get_stem_contributions_d3(vrna_fold_compound_t       *fc,
 #endif
 
   stems = (int *)vrna_alloc(sizeof(int) * (maxdist + 6));
-  stems -= i;
+  stems -= (size_t)i;
 
   sc_spl_stem = sc_wrapper->decomp_stem;
   sc_red_stem = sc_wrapper->red_stem;
@@ -1896,7 +1896,7 @@ f3_get_stem_contributions_d5(vrna_fold_compound_t       *fc,
 #endif
 
   stems = (int *)vrna_alloc(sizeof(int) * (maxdist + 6));
-  stems -= i;
+  stems -= (size_t)i;
 
   sc_spl_stem = sc_wrapper->decomp_stem1;
   sc_red_stem = sc_wrapper->red_stem;
@@ -2207,7 +2207,7 @@ f3_get_stem_contributions_d53(vrna_fold_compound_t      *fc,
 #endif
 
   stems = (int *)vrna_alloc(sizeof(int) * (maxdist + 6));
-  stems -= i;
+  stems -= (size_t)i;
 
   sc_spl_stem = sc_wrapper->decomp_stem1;
   sc_red_stem = sc_wrapper->red_stem;
